+++ src/main/dnacode.d
@@ -26,7 +26,7 @@ void main(string[] args){
     RNAstrand rna = RNAstrand_from_DNAstrand(dna);
     Protein[] p = Proteins_from_RNAstrand(rna);
     writefln("%s\n%s\n%s",dna,rna,p);
-    for(int x=0;x<1000;x++){
+    for(size_t x=0;x<1000;x++){
       dna = Elongate_DNA(dna);
       dna = Mutate_DNA(dna);
     }
